requirements.txt
twitter-api-client==0.10.22
python-dotenv==1.0.1
requests==2.32.3